objId is size_t
git-svn-id: file:///home/roberto/Git/svn@7140 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 22f656f278fbb4efa735f00ab1141685ee820a3e

+++ trunk/xmipp/libraries/reconstruction/angular_discrete_assign.cpp
@@ -653,7 +653,7 @@ int ProgAngularDiscreteAssign::pick_view(int method,
 // Run ---------------------------------------------------------------------
 // Predict shift and psi ---------------------------------------------------
 //#define DEBUG
-void ProgAngularDiscreteAssign::processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId)
+void ProgAngularDiscreteAssign::processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId)
 {
     // Read the image and take its angles from the Metadata
     // if they are available. If not, take them from the header.
